+++ R/zzz.R
@@ -17,7 +17,7 @@
       #While developing:
         packageStartupMessage(
          "#######################################################\n",
-                              "This Version 2026 04 28 - 06.12PM\n" )
+                              "This Version 2026 05 01 - 10.43AM\n" )